fix: 하루종일 이벤트 중 반복 일정 아이콘 표시
