+++ lib/home/components/day_view.dart
@@ -8,6 +8,8 @@ import 'package:intl/intl.dart';
 import 'package:lovendar/common/contants/layout_const.dart';
 import 'package:lovendar/common/schedule/model/schedule_model.dart';
 import 'package:lovendar/common/schedule/model/schedules_response.dart';
+import 'package:lovendar/home/components/schedule_container/all_day_schedule_box.dart';
+import 'package:lovendar/home/components/schedule_container/schedule_box.dart';
 import 'package:modal_bottom_sheet/modal_bottom_sheet.dart';
 import 'package:lovendar/common/schedule/provider/schedules_provider.dart';
 import 'package:lovendar/common/theme/theme_light.dart';
@@ -152,49 +154,16 @@ class _DayViewState extends ConsumerState<DayView> {
 
                           return Stack(
                             children: [
-                              // TODO: 여기에 전달 되는 event들은 특정 날짜에 해당하는 데이터여야 한다.
-                              // for (var scheduleGroup
-                              //     in groupOverlappingSchedules(
-                              //   schedules: schedules
-                              //       .where(
-                              //         (s) => s.isEventOnDate(
-                              //             targetDate: widget.date),
-                              //       )
-                              //       .toList()
-                              //       .adjustMultiDaySchedules(
-                              //           targetDate: widget.date),
-                              // ))
                               for (var scheduleGroup
                                   in groupOverlappingSchedules(
                                 schedules: scheduleObjs.adjustMultiDaySchedules(
                                     targetDate: widget.date),
                               ))
-                                // 수정 모드
-                                // for (var scheduleGroup
-                                //     in groupOverlappingSchedules(
-                                //   schedules: orgSchedules
-                                //       .where(
-                                //         (s) => s.scheduleData.isEventOnDate(
-                                //             targetDate: widget.date),
-                                //       )
-                                //       .toList()
-                                //       .map((s) => s.scheduleData)
-                                //       .toList()
-                                //       .adjustMultiDaySchedules(
-                                //           targetDate: widget.date),
-                                // ))
-                                // for (var scheduleGroup in schedules
-                                //     .where((s) =>
-                                //         s.isEventOnDate(targetDate: widget.date))
-                                //     .toList()
-                                // .groupOverlappingSchedules())
                                 for (var organizedSchedule
                                     in getOrganizedSchedules(
                                   overlappingSchedules: scheduleGroup,
                                   viewWidth: parentWidth,
                                 ))
-                                  // for (var organizedSchedule in scheduleGroup
-                                  //     .getOrganizedSchedules(parentWidth))
                                   ScheduleBox(
                                     organizedSchedule: organizedSchedule,
                                     targetDate: widget.date,
@@ -311,269 +280,6 @@ class _DayViewState extends ConsumerState<DayView> {
           ),
         ),
       ],
-    );
-  }
-}
-
-class AllDayScheduleBox extends ConsumerWidget {
-  final DateTime date;
-  final ScheduleType category;
-  final bool isTapped;
-  final void Function() toggleIsTapped;
-
-  const AllDayScheduleBox({
-    super.key,
-    required this.date,
-    required this.category,
-    required this.isTapped,
-    required this.toggleIsTapped,
-  });
-
-  @override
-  Widget build(BuildContext context, WidgetRef ref) {
-    // ScheduleModel? targetSchedule = scheduleObjs.firstWhere(
-    //     (s) => s.scheduleId == organizedSchedule.scheduleData.scheduleId);
-
-    final allDaySchedulesInDateByCat = ref
-        .read(schedulesProvider.notifier)
-        .getAllDaySchedulesByDateAndCat(date: date, category: category);
-
-    return Column(
-      children: [
-        for (int idx = 0; idx < allDaySchedulesInDateByCat.length; idx++)
-          // if (isTapped || idx == 0)
-          Padding(
-            padding: const EdgeInsets.only(bottom: 2.0),
-            child: GestureDetector(
-              onTap: () {
-                if (!isTapped) {
-                  if (isTapped || idx == 0) {
-                    toggleIsTapped();
-                  }
-                } else {
-                  showModalBottomSheet<void>(
-                    context: context,
-                    isScrollControlled: true,
-                    isDismissible: false,
-                    builder: (BuildContext context) {
-                      return SizedBox(
-                        height: MediaQuery.of(context).size.height * 0.85,
-                        child: ClipRRect(
-                          borderRadius: const BorderRadius.only(
-                            topLeft: Radius.circular(20),
-                            topRight: Radius.circular(20),
-                          ),
-                          child: ScheduleDetailScreen(
-                            date: date,
-                            intialSchedule: allDaySchedulesInDateByCat[idx],
-                          ),
-                        ),
-                      );
-                    },
-                  );
-                }
-              },
-              child: AnimatedContainer(
-                duration: Duration(milliseconds: 300),
-                curve: Curves.easeInOut,
-                height: isTapped || idx == 0 ? 27.5 : 0,
-                child: Container(
-                  alignment: Alignment.centerLeft,
-                  padding: EdgeInsets.symmetric(
-                    horizontal: 10,
-                    vertical: 3,
-                  ),
-                  height: 27.5,
-                  width: double.infinity,
-                  decoration: BoxDecoration(
-                    color: allDaySchedulesInDateByCat[idx].color,
-                    borderRadius: BorderRadius.circular(4),
-                    boxShadow: [
-                      BoxShadow(
-                          color: Colors.grey.withOpacity(0.1),
-                          spreadRadius: 0,
-                          blurRadius: 10,
-                          offset: Offset.fromDirection(
-                              360, 10) // changes position of shadow
-                          ),
-                    ],
-                  ),
-                  child: Row(
-                    mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                    children: [
-                      Text(
-                        allDaySchedulesInDateByCat[idx].title,
-                        style: TextStyle(
-                          fontSize: 11.5,
-                          fontWeight: FontWeight.bold,
-                        ),
-                      ),
-                      allDaySchedulesInDateByCat.length > 1 && !isTapped
-                          ? Text(
-                              "+${allDaySchedulesInDateByCat.length - 1}개 더보기",
-                              style: TextStyle(
-                                fontSize: 10,
-                                fontWeight: FontWeight.w500,
-                                color: Colors.grey[600]!,
-                              ),
-                            )
-                          : SizedBox.shrink(),
-                    ],
-                  ),
-                ),
-              ),
-            ),
-          ),
-      ],
-    );
-  }
-}
-
-class ScheduleBox extends ConsumerWidget {
-  final OrganizedSchedule organizedSchedule;
-  final DateTime targetDate;
-
-  const ScheduleBox({
-    super.key,
-    required this.organizedSchedule,
-    required this.targetDate,
-  });
-
-  @override
-  Widget build(BuildContext context, WidgetRef ref) {
-    final schedules = ref.watch(schedulesProvider);
-    final scheduleObjs = schedules
-        .map(
-          (s) => s.copyScheduleOnDate(targetDate: targetDate),
-        )
-        .whereType<ScheduleModel>()
-        .toList();
-
-    ScheduleModel? targetSchedule = scheduleObjs.firstWhere(
-        (s) => s.scheduleId == organizedSchedule.scheduleData.scheduleId);
-
-    return Positioned(
-      left: organizedSchedule.left,
-      top: organizedSchedule.top,
-      width: organizedSchedule.width - LayoutConst.scheduleBoxGap,
-      height: organizedSchedule.height,
-      child: ClipRRect(
-        borderRadius: BorderRadius.circular(6),
-        child: Material(
-          color: organizedSchedule.scheduleData.color,
-          // shape: RoundedRectangleBorder(
-          //   borderRadius: BorderRadius.circular(6),
-          // ),
-          child: InkWell(
-            onTap: () {
-              showModalBottomSheet(
-                context: context,
-                isScrollControlled: true,
-                builder: (context) => SizedBox(
-                  height: MediaQuery.of(context).size.height * 0.85,
-                  child: ClipRRect(
-                    borderRadius: const BorderRadius.only(
-                      topLeft: Radius.circular(20),
-                      topRight: Radius.circular(20),
-                    ),
-                    child: ScheduleDetailScreen(
-                      date: targetDate,
-                      intialSchedule:
-                          targetSchedule ?? organizedSchedule.scheduleData,
-                    ),
-                  ),
-                ),
-              );
-            },
-            splashColor: AppTheme.getDarkerColorByScheduleType(
-                    organizedSchedule.scheduleData.category)
-                .withOpacity(0.2), // 리플 색상
-            child: Container(
-              padding: EdgeInsets.symmetric(
-                vertical: max(organizedSchedule.height * 0.06, 7),
-                horizontal: max(organizedSchedule.width * 0.05, 5),
-              ),
-              width: organizedSchedule.width,
-              height: organizedSchedule.height,
-              decoration: BoxDecoration(
-                boxShadow: [
-                  BoxShadow(
-                    color: Colors.grey.withOpacity(0.1),
-                    spreadRadius: 0,
-                    blurRadius: 10,
-                    offset: Offset.fromDirection(
-                        360, 10), // changes position of shadow
-                  ),
-                ],
-              ),
-              child: SingleChildScrollView(
-                child: Column(
-                  crossAxisAlignment: CrossAxisAlignment.start,
-                  children: [
-                    Wrap(
-                      spacing: 4,
-                      alignment: WrapAlignment.spaceBetween,
-                      crossAxisAlignment: WrapCrossAlignment.center,
-                      children: [
-                        if (organizedSchedule.scheduleData.isRepeat)
-                          Icon(
-                            Icons.replay_rounded,
-                            size: 13,
-                          ),
-                        Text(
-                          organizedSchedule.scheduleData.title,
-                          maxLines: 1,
-                          style: const TextStyle(
-                            fontSize: 11.5,
-                            fontWeight: FontWeight.bold,
-                          ),
-                          overflow: TextOverflow.ellipsis,
-                        ),
-                      ],
-                    ),
-
-                    Wrap(
-                      children: [
-                        Text(
-                          "${DateFormat("h시 m분").format(organizedSchedule.scheduleData.startTime)} ",
-                          maxLines: 1,
-                          overflow: TextOverflow.ellipsis,
-                          style: TextStyle(
-                            fontSize: 9,
-                            fontWeight: FontWeight.w500,
-                            color: Colors.grey[600]!,
-                          ),
-                        ),
-                        Text(
-                          "· ${organizedSchedule.scheduleData.duration.inHours}h"
-                              .toString(),
-                          maxLines: 1,
-                          overflow: TextOverflow.ellipsis,
-                          style: TextStyle(
-                            fontSize: 9,
-                            fontWeight: FontWeight.w500,
-                            color: Colors.grey[600]!,
-                          ),
-                        ),
-                      ],
-                    ),
-                    // Text(
-                    //   "${DateFormat.jm().format(organizedSchedule.scheduleData.startTime)} · ${organizedSchedule.scheduleData.duration.inHours}h",
-                    //   maxLines: 1,
-                    //   overflow: TextOverflow.ellipsis,
-                    //   style: TextStyle(
-                    //     fontSize: 10,
-                    //     fontWeight: FontWeight.w500,
-                    //     color: Colors.grey[600]!,
-                    //   ),
-                    // ),
-                  ],
-                ),
-              ),
-            ),
-          ),
-        ),
-      ),
     );
   }
 }
